warningのdisableを削除 警告していたところを修正

+++ Src/BaseProc.cpp
@@ -215,7 +215,7 @@ CBaseObj::CBaseObj(CBaseProc* pProc)
 	ResetStatus();						// 各種ステータスをリセット
 	m_bActive = false;					// true:表示  false:非表示
 	m_dwObjID = pProc->GetProcID();		// オブジェクトＩＤ
-	m_dwObjNo = pProc->GetObjArrayPtr().size();	// オブジェクトＮＯ  プッシュバック前の配列サイズがオブジェクトＮＯとなる
+	m_dwObjNo = (DWORD)pProc->GetObjArrayPtr().size();	// オブジェクトＮＯ  プッシュバック前の配列サイズがオブジェクトＮＯとなる
 	m_nMeshIdx = 0;						// メッシュ配列の添字           // -- 2022.12.20
 	m_pBBox = nullptr;						// バウンディングボックスクラス
 	m_mWorld = XMMatrixIdentity();		// ワールドマトリクス
